fix: Move base test class into tests dir
