+++ tests/__init__.py
@@ -1,0 +1,4 @@
+from django import setup
+import os
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "core.settings")
+setup()
